+++ OpenGL/src/Graphics/GLs/IndexBuffer.cpp
@@ -3,7 +3,7 @@
 #include "GLDebug.h"
 
 namespace Graphics {
-    IndexBuffer::IndexBuffer(std::span<const uint> data) : GLObject({}), length(data.size()) {
+    IndexBuffer::IndexBuffer(std::span<const uint> data) : GLObject({}), length((uint)data.size()) {
         GL_CALL(glBufferData(GL_ELEMENT_ARRAY_BUFFER, data.size_bytes(), data.data(), GL_STATIC_DRAW));
     }
 }
